// next-app/src/lib/types.ts
export interface KhachHangModel {
  id_KhachHang: string;
  tenKhachHang: string | null;
  email: string | null;
  soDienThoai: string | null;
  diaChi: string | null;
  diemThuong: number | null;
}